deployment-manager/functions/pubsub_to_bigquery/requirements.txt
google-cloud-bigquery>=0.27
